added cleanup and documentation
SourceMap moved to its own file to reduce the size
of index.rs
